ref (recipe page): Move prefetch to Page Container for better maintainability.
