+++ src/app/@modal/(.)sn/recipes/[recipeKey]/page.tsx
@@ -1,11 +1,3 @@
-import {
-  dehydrate,
-  HydrationBoundary,
-  QueryClient,
-} from '@tanstack/react-query';
-
-import { recipeOptions, recipeRecommendOptions } from '@/queries/options';
-
 import ModalRecipe from './ModalRecipe';
 
 interface Props {
@@ -16,17 +8,7 @@ async function RecipePage({ params }: Props) {
   const { recipeKey } = await params;
   if (!recipeKey) return null;
 
-  const queryClient = new QueryClient();
-  await Promise.all([
-    queryClient.prefetchQuery(recipeOptions({ key: recipeKey })),
-    queryClient.prefetchQuery(recipeRecommendOptions({ key: recipeKey })),
-  ]);
-
-  return (
-    <HydrationBoundary state={dehydrate(queryClient)}>
-      <ModalRecipe />
-    </HydrationBoundary>
-  );
+  return <ModalRecipe recipeKey={recipeKey} />;
 }
 
 export default RecipePage;
