+++ src/app/sn/recipes/[recipeKey]/Recipe/style.module.scss
@@ -53,12 +53,17 @@
   }
 }
 
+.border {
+  margin: 0 20px;
+  border-left: 1px solid $black;
+}
+
 .content {
   flex: 1;
   display: flex;
   flex-direction: column;
   align-items: center;
-  border-left: 1px solid;
+
   overflow: overlay;
 
   @media #{$laptop} {
