+++ src/app/sn/recipes/[recipeKey]/RecipePageContainer.tsx
@@ -1,28 +1,29 @@
-'use client';
+import {
+  dehydrate,
+  HydrationBoundary,
+  QueryClient,
+} from '@tanstack/react-query';
 
-import { notFound, useParams } from 'next/navigation';
-import { useQuery } from '@tanstack/react-query';
-
-import { recipeOptions } from '@/queries/options';
+import { recipeOptions, recipeRecommendOptions } from '@/queries/options';
 
 import Recipe from './Recipe';
 import Recommend from './Recommend';
-import { RecipePageParams } from './page';
 
-function RecipePageContainer() {
-  const params = useParams<RecipePageParams>();
-  const { data: recipe, isError } = useQuery(
-    recipeOptions({ key: params.recipeKey }),
-  );
+async function RecipePageContainer({ recipeKey }: { recipeKey: string }) {
+  const queryClient = new QueryClient();
 
-  if (isError) throw new Error('Failed to load recipe');
-  if (!recipe) return notFound();
+  await Promise.all([
+    queryClient.prefetchQuery(recipeOptions({ key: recipeKey })),
+    queryClient.prefetchQuery(recipeRecommendOptions({ key: recipeKey })),
+  ]);
 
   return (
-    <div>
-      <Recipe recipe={recipe} />
-      <Recommend />
-    </div>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <div>
+        <Recipe />
+        <Recommend />
+      </div>
+    </HydrationBoundary>
   );
 }
